Fixed menus
fixing scores and leadership scripts
