+++ BreakoutNextLevel/Assets/Scenes/Profile Info/profileInfo.unity
@@ -154,11 +154,11 @@ RectTransform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1676294225}
-  m_RootOrder: 5
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -6, y: -123.651505}
+  m_AnchoredPosition: {x: -6, y: -88.00001}
   m_SizeDelta: {x: 380.7166, y: 76.583}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &70191027
@@ -234,11 +234,11 @@ RectTransform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1676294225}
-  m_RootOrder: 3
+  m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0.0000047684, y: 27.999992}
+  m_AnchoredPosition: {x: 0.0000047684, y: -13.000014}
   m_SizeDelta: {x: 352.05, y: 75.28}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &116272043
@@ -380,90 +380,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bb43a3d5ba595f3478824eb2b547f29d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  ProfileNameText: {fileID: 568228664}
+  coinsToSave: 0
+  skinsToSave: 0
+  levelsToSave: 0
+  ProfileNameText: {fileID: 1759756169}
   CoinText: {fileID: 116272043}
-  LevelsCompletedText: {fileID: 1759756169}
   SkinText: {fileID: 70191027}
---- !u!1 &568228662
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 568228663}
-  - component: {fileID: 568228665}
-  - component: {fileID: 568228664}
-  m_Layer: 0
-  m_Name: ProfileNameText
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!224 &568228663
-RectTransform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 568228662}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 1676294225}
-  m_RootOrder: 2
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.5, y: 0.5}
-  m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0.0000047684, y: 124.78}
-  m_SizeDelta: {x: 352.05, y: 75.28}
-  m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &568228664
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 568228662}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5f7201a12d95ffc409449d95f23cf332, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_Material: {fileID: 0}
-  m_Color: {r: 0.40926296, g: 0.3952296, b: 0.49056602, a: 1}
-  m_RaycastTarget: 1
-  m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
-  m_Maskable: 1
-  m_OnCullStateChanged:
-    m_PersistentCalls:
-      m_Calls: []
-  m_FontData:
-    m_Font: {fileID: 12800000, guid: 997a43b767814dd0a7642ec9b78cba41, type: 3}
-    m_FontSize: 35
-    m_FontStyle: 0
-    m_BestFit: 0
-    m_MinSize: 2
-    m_MaxSize: 45
-    m_Alignment: 7
-    m_AlignByGeometry: 0
-    m_RichText: 1
-    m_HorizontalOverflow: 0
-    m_VerticalOverflow: 0
-    m_LineSpacing: 1
-  m_Text: 
---- !u!222 &568228665
-CanvasRenderer:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 568228662}
-  m_CullTransparentMesh: 1
 --- !u!1 &992947665
 GameObject:
   m_ObjectHideFlags: 0
@@ -565,7 +487,6 @@ RectTransform:
   m_Children:
   - {fileID: 2103318178}
   - {fileID: 1790633382}
-  - {fileID: 568228663}
   - {fileID: 116272042}
   - {fileID: 1759756168}
   - {fileID: 70191026}
@@ -651,7 +572,7 @@ GameObject:
   - component: {fileID: 1759756170}
   - component: {fileID: 1759756169}
   m_Layer: 0
-  m_Name: LevelsCompleteText
+  m_Name: ProfileNameText
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -670,11 +591,11 @@ RectTransform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1676294225}
-  m_RootOrder: 4
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -7, y: -48.36001}
+  m_AnchoredPosition: {x: -7, y: 63}
   m_SizeDelta: {x: 396.3531, y: 74}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1759756169
@@ -827,7 +748,7 @@ RectTransform:
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1676294225}
-  m_RootOrder: 6
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
   m_AnchorMax: {x: 0, y: 1}
