--- BreakoutNextLevel/Assets/Scripts/LeaderBoard.cs
@@ -1,20 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class Scores : MonoBehaviour
-{
-    void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Escape))
-        {
-            SceneManager.LoadScene("MainMenu");
-        }
-    }
-
-    public void goHome()
-    {
-        SceneManager.LoadScene("MainMenu");
-    }
-}
