--- BreakoutNextLevel/Assets/Scripts/LeaderBoard.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 374e80e12fdd94a43ba9c2bda6b66bb5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
